+++ axum-macros/CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- Add `#[derive(TypedPath)]` for use with axum-extra's new "type safe" routing API ([#756])
+- **added:** Add `#[derive(TypedPath)]` for use with axum-extra's new "type safe" routing API ([#756])
+- **added:** `#[derive(TypedPath)]` now also generates a `TryFrom<_> for Uri`
+  implementation ([#790])
+
+[#790]: https://github.com/tokio-rs/axum/pull/790
 
 # 0.1.0 (31. January, 2022)